+++ display/describe_parse_error.js
@@ -1,0 +1,16 @@
+const {parseErrors} = require('./constants');
+
+const defaultError = '#ERROR!';
+
+/** Describe a hot formula parser error
+
+  {
+    error: <Error Type String>
+  }
+
+  @returns
+  <Display String>
+*/
+module.exports = ({error}) => {
+  return parseErrors[error] || parseErrors[defaultError];
+};