+++ display/index.js
@@ -1,5 +1,6 @@
 const chartAliasForPeer = require('./chart_alias_for_peer');
 const describeConfidence = require('./describe_confidence');
+const describeParseError = require('./describe_parse_error');
 const describeRoute = require('./describe_route');
 const describeRoutingFailure = require('./describe_routing_failure');
 const formatFeeRate = require('./format_fee_rate');
@@ -11,6 +12,7 @@ const sumsForSegment = require('./sums_for_segment');
 module.exports = {
   chartAliasForPeer,
   describeConfidence,
+  describeParseError,
   describeRoute,
   describeRoutingFailure,
   formatFeeRate,