+++ network/probe_destination.js
@@ -30,6 +30,7 @@ const fromKeyType = '34349339';
 const {isArray} = Array;
 const keySendPreimageType = '5482373484';
 const makeNonce = () => randomBytes(32).toString('hex');
+const {max} = Math;
 const messageType = '34349334';
 const nodeKeyFamily = 6;
 const preimageByteLength = 32;
@@ -374,8 +375,14 @@ module.exports = (args, cbk) => {
         });
 
         sub.once('error', err => cbk(err));
-        sub.once('failure', () => cbk(null, {maximum: Number()}));
-        sub.once('success', ({maximum}) => cbk(null, {maximum}));
+
+        // Did not find any higher value routes
+        sub.once('failure', () => cbk(null, {maximum: probe.route.tokens}));
+
+        // Found a successful high value route
+        sub.once('success', ({maximum}) => {
+          return cbk(null, {maximum: max(maximum, probe.route.tokens)});
+        });
 
         return;
       }],