+++ swaps/index.js
@@ -1,6 +1,7 @@
 const getPaidService = require('./get_paid_service');
 const getSwapCost = require('./get_swap_cost');
 const getSwapService = require('./get_swap_service');
+const manageRebalance = require('./manage_rebalance');
 const rebalance = require('./rebalance');
 const swapApiKey = require('./swap_api_key');
 const swapIn = require('./swap_in');
@@ -11,6 +12,7 @@ module.exports = {
   getPaidService,
   getSwapCost,
   getSwapService,
+  manageRebalance,
   rebalance,
   swapApiKey,
   swapIn,